criação do projeto de cliente da api, com buefy e criação da tela de listagem de gêneros

+++ database/migrations/2020_04_22_211916_create_movies_table.php
@@ -16,7 +16,7 @@ class CreateMoviesTable extends Migration
         Schema::create('movies', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->tinyInteger('year');
+            $table->unsignedInteger('year');
             $table->string('trailer_url');
             $table->string('description');
             $table->json('metadata');
